Move Kickstart build methods into their own module
As their are now two Kickstart build methods, they have been moved
into a dedicated module. Due to ruby naming the conventions, the Base
class has became Kickstart and the old Kickstart method is now Legacy.
