+++ src/build_methods.rb
@@ -3,6 +3,7 @@
 
 require 'utils/dynamic_require'
 require 'build_methods/build_method'
+require 'build_methods/kickstarts/kickstart'
 
 Metalware::Utils::DynamicRequire.relative('build_methods')
