+++ src/node.rb
@@ -186,11 +186,11 @@ module Metalware
     def build_method_class
       case repo_config[:build_method]&.to_sym
       when :uefi
-        BuildMethods::UEFI
+        BuildMethods::Kickstarts::UEFI
       when :basic
         BuildMethods::Basic
       else
-        BuildMethods::Kickstart
+        BuildMethods::Kickstarts::Legacy
       end
     end
   end
